Add readme.md (#19)
* Initial introduction and headers

Placeholders mainly.

* WIP

checking it current working version to be

* Simplified index

Simplifying the index and handing off to other documents.

* Removing gumpf.

Cutting out a lot of gumpf

* mv index.md > readme.md and expand

Adds initial draft of ready with topic based guidance for

* why we’re doing this
* how to contribute
* project overview

and a note about improving documentation, which I will add to all other
documents.

Symlinks index.md for gh-pages into readme.md. as a bit of a test.